Use `dtype="float64"` in `pd.Series` to address FutureWarning

+++ sktime/forecasting/base/adapters/_pmdarima.py
@@ -109,7 +109,7 @@ class _PmdArimaAdapter(BaseForecaster):
         # Initialize return objects
         fh_abs = fh.to_absolute(self.cutoff).to_numpy()
         fh_idx = fh.to_indexer(self.cutoff, from_cutoff=False)
-        y_pred = pd.Series(index=fh_abs)
+        y_pred = pd.Series(index=fh_abs, dtype="float64")
 
         # for in-sample predictions, pmdarima requires zero-based integer indicies
         start, end = fh.to_absolute_int(self._y.index[0], self.cutoff)[[0, -1]]
